% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaData.R
\name{AddMetaData}
\alias{AddMetaData}
\title{AddMetaData
This function adds metadata to the original matrix. It can include up to one parameter and one batch identifiers.}
\usage{
AddMetaData(MetaDataObject, BatchFrame = NULL, BatchName = NULL,
  ParameterFrame = NULL, ParameterName = NULL)
}
\arguments{
\item{BatchFrame}{Character vector of names corresponding to batches in sample extraction.}

\item{BatchName}{Character string of batch.}

\item{ParameterFrame}{Vector corresponding to information about samples.}

\item{ParameterName}{Character string to name parameter}

\item{SampleVector}{Character vector of sample names.}
}
\description{
AddMetaData
This function adds metadata to the original matrix. It can include up to one parameter and one batch identifiers.
}
\examples{
AddMetaData()

}
\keyword{addmeta}
