% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentStatisticsforTax.R
\name{AlignmentStatisticsforTax}
\alias{AlignmentStatisticsforTax}
\title{AlignmentStatisticsforTax
This function generates alignment information and statistics for a specific microbe at taxonomic level.}
\usage{
AlignmentStatisticsforTax(sample, tax.id, tax.level = "species",
  bin.size = 10^3, Merge.data.frame)
}
\arguments{
\item{tax.id}{ncbi taxon identifier}

\item{tax.level}{ncbi taxification level. "species" by DEFAULT.}

\item{bin.size}{Size of bin in bp. Default is 1000 bp.}

\item{Merge.data.frame}{Data frame including alignment and abundance information}

\item{Sample}{Abundance matrix to be loaded. Must be filtered to tri-column form.}
}
\description{
AlignmentStatisticsforTax
This function generates alignment information and statistics for a specific microbe at taxonomic level.
}
\examples{
AlignmentStatisticsforTax()

}
\keyword{stats}
