% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignStatsSample.R
\name{AlignStatsSample}
\alias{AlignStatsSample}
\title{AlignStatsSample
This function generates alignment information and statistics for a specific microbe at taxonomic level.}
\usage{
AlignStatsSample(Sample, BinSize = 10^3, TblatPath = "./",
  GITable = "gi_tax_info.tab", MinHits = 5, OutPath = "./",
  Overwrite = F)
}
\arguments{
\item{Sample}{Sample name to load in tblat.1 file}

\item{BinSize}{Size of bin in bp. Default is 1000 bp. Numeric. 1000 bp by default.}

\item{TblatPath}{Path to tblat file directory. Character. Current directory by default.}

\item{GITable}{GI/Tax table path. Character.}

\item{MinHits}{Minimum number of hits to consider. Numeric. 5 by default.}

\item{OutPath}{Path to write out tables.}

\item{Overwrite}{Overwrite existing file. Logical. FALSE by default.}

\item{Tax}{ID number for taxon.}
}
\description{
AlignStatsSample
This function generates alignment information and statistics for a specific microbe at taxonomic level.
}
\examples{
AlignStatsSample()

}
\keyword{stats}
