% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignStatsTax.R
\name{AlignStatsTax}
\alias{AlignStatsTax}
\title{AlignStatsTax
This function generates alignment information and statistics for a specific microbe at taxonomic level.}
\usage{
AlignStatsTax(Sample, Tax, BinSize = 10^3, TblatGIObject)
}
\arguments{
\item{Sample}{Sample name to load in tblat.1 file}

\item{Tax}{ID number for taxon.}

\item{BinSize}{Size of bin in bp. Default is 1000 bp. Numeric. 1000 bp by default.}

\item{TblatGIObject}{Tblat / GI/Tax table path. Character.}
}
\description{
AlignStatsTax
This function generates alignment information and statistics for a specific microbe at taxonomic level.
}
\examples{
AlignStatsTax()

}
\keyword{stats}
