% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchAssemble.R
\name{BatchAssemble}
\alias{BatchAssemble}
\title{BatchAssemble
This function is used to assemble a dataframe containing the variance of each taxa in samples.}
\usage{
BatchAssemble(BatchList, metric = "variation")
}
\arguments{
\item{BatchList}{List of batch objects.}

\item{metric}{"mean" or "variation". "variation" by DEFAULT.}
}
\description{
BatchAssemble
This function is used to assemble a dataframe containing the variance of each taxa in samples.
}
\examples{
BatchAssemble()

}
\keyword{batch}
