% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryScore.R
\name{BinaryScore}
\alias{BinaryScore}
\title{BinaryScore
This function determines a score from linear combination of weights with respect to TP, TN, FP, FN, and untested counts.}
\usage{
BinaryScore(AbundanceObject, MetaDataObject, AbundanceTaxName,
  MetaDataTaxName, NegativeName = "Negative",
  CIBacteriaVector = c("Escherichia", "Enterococcus"),
  WeightsVector = c(1, 2, -1, -3, -0.1))
}
\arguments{
\item{AbundanceObject}{Subset object with abundance}

\item{MetaDataObject}{Metadata object containing biomass.}

\item{AbundanceTaxName}{Character value for vector name of clinical taxa identified.}

\item{MetaDataTaxName}{Character value for vector name of metagenomic taxa identified.}

\item{CIBacteriaVector}{Character vector of taxa tested. Enterococcus and Escherichia genera by DEFAULT.}

\item{WeightsVector}{Numeric vector of weights for (in order): true positive, true negative, false postive, false negative, and untested. DEFAULT vector provided.}

\item{MetaDataTaxName}{String to identify negative tests. "Negative" by DEFAULT.}
}
\description{
BinaryScore
This function determines a score from linear combination of weights with respect to TP, TN, FP, FN, and untested counts.
}
\examples{
BinaryScore()

}
\keyword{score}
