% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomassCorrelation.R
\name{BiomassCorrelation}
\alias{BiomassCorrelation}
\title{BiomassCorrelation
This function looks for correlation between a particular taxon abundance and biomass.
For negatively correlated taxa, biomass-proportions are transformed and scaled.}
\usage{
BiomassCorrelation(AbundanceObject, MetaDataObject,
  BiomassVariable = "Biomass", Cor.signif.threshold = 10^-2,
  Zscore.threshold = 1.65)
}
\arguments{
\item{AbundanceObject}{Subset object with abundance}

\item{MetaDataObject}{Metadata object containing biomass.}

\item{BiomassVariable}{string to describe column name of biomass vector in metadata frame. "Biomass" by DEFAULT.}

\item{Cor.signif.threshold}{Maximum value for p value of correlated used to initialize Box-cox. 10**-2 by DEFAULT.}

\item{Zscore.threshold}{Maximum threshold to reject values. 1.65 by DEFAULT (roughly 95%).}
}
\description{
BiomassCorrelation
This function looks for correlation between a particular taxon abundance and biomass.
For negatively correlated taxa, biomass-proportions are transformed and scaled.
}
\examples{
BiomassCorrelation()

}
\keyword{biomass}
