% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToMatrix.R
\name{ConvertToMatrix}
\alias{ConvertToMatrix}
\title{ConvertToMatrix
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.}
\usage{
ConvertToMatrix(AbundanceObject, CastingValue, Samples.filter = NULL,
  Taxon.filter = NULL)
}
\arguments{
\item{AbundanceObject}{Abundance matrix to be loaded. Must be filtered to tri-column form.}

\item{CastingValue}{Character value, reflecting column name, to fill in matrix. "Measurement" is DEFAULT.}

\item{Samples.filter}{Character vector of samples to remove. NULL by DEFAULT.}

\item{Taxon.filter}{Character vector of taxa to remove. NULL by DEFAULT.}
}
\description{
ConvertToMatrix
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.
}
\examples{
ConvertToMatrix()
}
\keyword{convert}
