% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenoiseAlgorithm.R
\name{DenoiseAlgorithm}
\alias{DenoiseAlgorithm}
\title{DenoiseAlgorithm
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.}
\usage{
DenoiseAlgorithm(AbundanceObject, MetaDataObject, NegativeObject,
  ReadAbundMatrix, CV.Filter = T, MassVar.Filter = T,
  NegCtrl.Filter = T, deltaCV.Param = 10^5, MassVar.Param = 0,
  NegCtrl.Param = 0, TaxLevel = "genus", FastqPath = "./",
  TblatPath = "./", Measurement = "RelCoverage", ReturnContams = F,
  GITable = "gi_tax_info.tab")
}
\arguments{
\item{AbundanceObject}{Matrix including BLAST hits for each sample.}

\item{MetaDataObject}{Metadata matrix to describe biomass and batches.}

\item{NegativeObject}{Negative control matrix.}

\item{ReadAbundMatrix}{Matrix listing samples and read abundance.}

\item{CV.Filter}{Filter matrix based on CV. Logical. TRUE by default.}

\item{MassVar.Filter}{Filter matrix based on mass variation. Logical. TRUE by default.}

\item{NegCtrl.Filter}{Filter matrix based on negative control taxa. Logical. TRUE by default.}

\item{deltaCV.Param}{CV based parameter for filtering difference (max value). Numeric. 10**5 by default.}

\item{MassVar.Param}{Mass variation based parameter for filtering (min value). Numeric. 0 by default.}

\item{NegCtrl.Param}{Negative control based parameter for filtering (min value). Numeric. 0 by default.}

\item{TaxLevel}{Taxonomic level to perform analysis. Character. "species" by default.}

\item{FastqPath}{Path to raw, gzipped fastqs. Current directory by default.}

\item{TblatPath}{Path to tblat files. Current directory by default.}

\item{Measurement}{Name of column in abundance object containing measurements. Character. "RelCoverage" by default.}

\item{ReturnContams}{Return contaminants instead of filtered matrix. Logical. FALSE by default.}
}
\description{
DenoiseAlgorithm
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.
}
\examples{
DenoiseAlgorithm()
}
\keyword{denoise}
