% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterWithNegativeControl.R
\name{FilterWithNegativeControl}
\alias{FilterWithNegativeControl}
\title{FilterWithNegativeControl
This function is used to find the proportion of all taxa present in the negative controls.}
\usage{
FilterWithNegativeControl(sample, contam.set, out.path = "./",
  factor.contam = 10, filter.out.list = T, raw.data.path, tblat.path)
}
\arguments{
\item{sample}{Sample name. Character.}

\item{contam.set}{Data frame of taxa in the contaminant samples.}

\item{out.path}{Pathway to store total reads. Character.}

\item{factor.contam}{Multiplier to set limits for contaminant taxa. Numeric.}

\item{filter.out.list}{Determines if the output is filtered out taxa (TRUE) or filtered in taxa (FLASE). Logical. TRUE by default.}

\item{raw.data.path}{Pathway to find zipped fastq files. Character.}

\item{tblat.path}{Pathway to alignment files (tblat.1 format). Character.}
}
\description{
FilterWithNegativeControl
This function is used to find the proportion of all taxa present in the negative controls.
}
\examples{
FilterWithNegativeControl()
}
\keyword{filter}
