% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateStats.R
\name{GenerateStats}
\alias{GenerateStats}
\title{GenerateStats
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.}
\usage{
GenerateStats(AbundanceObject)
}
\arguments{
\item{AbundanceObject}{Abundance matrix to be loaded. Must be filtered to tri-column form.}
}
\description{
GenerateStats
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.
}
\examples{
GenerateStats()
}
\keyword{stats}
