% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeMetaData.R
\name{InitializeMetaData}
\alias{InitializeMetaData}
\title{InitializeMetaData
This function initializes metadata matrix. It should at least include batch information.}
\usage{
InitializeMetaData(SampleVector, BatchVector, BatchName)
}
\arguments{
\item{SampleVector}{Character vector of sample names.}

\item{BatchVector}{Character vector of names corresponding to batches in sample extraction.}

\item{BatchName}{Character string of batch.}
}
\description{
InitializeMetaData
This function initializes metadata matrix. It should at least include batch information.
}
\examples{
InitializeMetaData()

}
\keyword{meta}
