% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAbundance.R
\name{LoadAbundance}
\alias{LoadAbundance}
\title{LoadAbundance}
\usage{
LoadAbundance(dir = "./", file)
}
\arguments{
\item{dir}{Directory containing microbial abundance matrix. DEFAULT is current directory.}

\item{file}{Abundance matrix filename.}
}
\description{
This function reads in an abundance matrix. Matrix should have rows corresponding to samples and microbes, e.g. row 1: Sample #1, Escherichia coli, 0.78
}
\examples{
LoadAbundance()
}
\keyword{load}
