% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCorrelation.R
\name{MetaCorrelation}
\alias{MetaCorrelation}
\title{MetaCorrelation
This function looks for correlation between a particular taxon abundance and biomass.}
\usage{
MetaCorrelation(tax.ID, AbundanceObject, AbundanceVariable, MetaDataObject,
  FeatureVariable)
}
\arguments{
\item{tax.ID}{NCBI taxification identifier}

\item{AbundanceObject}{Subset object with abundance}

\item{AbundanceVariable}{Character value on which to produce correlation.}

\item{MetaDataObject}{Metadata object containing biomass.}

\item{BiomassVariable}{Character value for biomass column.}

\item{tax.ID}{NCBI taxification identifier.}
}
\description{
MetaCorrelation
This function looks for correlation between a particular taxon abundance and biomass.
}
\examples{
MetaCorrelation()

}
\keyword{biomass}
