% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetAbundance.R
\name{SubsetAbundance}
\alias{SubsetAbundance}
\title{SubsetAbundance
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.}
\usage{
SubsetAbundance(AbundanceObject, tax.level = "species", measurement,
  sample.column, include.samples = c(), exclude.samples = c(),
  proportional.abundance = T)
}
\arguments{
\item{AbundanceObject}{Abundance matrix to be loaded.}

\item{tax.level}{Column name corresponding to taxification on which data will be aggregated. "species" by default.}

\item{measurement}{Column name corresponding to abundance measurement.}

\item{sample.column}{Column number corresponding to sample.}

\item{include.samples}{Names of samples to be included for subsetting.}

\item{exclude.samples}{Names of samples to be excluded for subsetting.}

\item{proportional.abundance}{Determine the abundance with respect to other measurements within the same Sample. TRUE by default.}
}
\description{
SubsetAbundance
This function allows specification of taxon level if applicable and will sum data to that level, depending on specific column being measured.
}
\examples{
SubsetAbundance()
}
\keyword{subset}
