% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalReadsGen.R
\name{TotalReadsGen}
\alias{TotalReadsGen}
\title{TotalReadsGen
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.}
\usage{
TotalReadsGen(MetaDataObject, TotalReadsOutput = NA, RawDataPath = NA)
}
\arguments{
\item{MetaDataObject}{Metadata matrix to describe biomass and batches.}

\item{TotalReadsOutput}{Path and filename to store TotalReadsFrame. Character. NA by default.}

\item{RawDataPath}{Path where zipped fastq files reside. Character. NA by default.}
}
\description{
TotalReadsGen
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.
}
\examples{
TotalReadsGen()
}
\keyword{totalreads}
