% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithinBatchVariation.R
\name{WithinBatchVariation}
\alias{WithinBatchVariation}
\title{WithinBatchVariation
This function looks for correlation between a particular taxon abundance and biomass.}
\usage{
WithinBatchVariation(AbundanceObject, AbundanceVariable, MetaDataObject,
  BiomassInputAdjust = F, BiomassVariable = NULL, BatchName = NULL,
  BatchID = NULL, Correlate.threshold = 10^-4)
}
\arguments{
\item{AbundanceObject}{Subset object with abundance}

\item{AbundanceVariable}{Character value on which to produce correlation.}

\item{MetaDataObject}{Metadata object containing biomass.}

\item{BiomassInputAdjust}{Boolean value to declare adjustment by biomass.FALSE by DEFAULT}

\item{BiomassVariable}{string to describe column name of biomass vector in metadata frame. NULL by DEFAULT}

\item{BatchName}{Character value for Batch column. NULL by DEFAULT.}

\item{BatchID}{Character value for specific batch. NULL by DEFAULT.}

\item{Correlate.threshold}{Float value of lowerbound of correlation between taxa. 10**-4 by DEFAULT}
}
\description{
WithinBatchVariation
This function looks for correlation between a particular taxon abundance and biomass.
}
\examples{
WithinBatchVariation()

}
\keyword{variance}
