% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadProportions.R
\name{ReadProportions}
\alias{ReadProportions}
\title{ReadProportions
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.}
\usage{
ReadProportions(sample, out.path = "./", raw.data.path, tblat.path,
  ReadAbundMatrix = c())
}
\arguments{
\item{sample}{Sample name. Character.}

\item{out.path}{Pathway to .totalreads.tab file. Current directory by default.}

\item{raw.data.path}{Pathway to find zipped fastq files. Character.}

\item{tblat.path}{Pathway to alignment files (tblat.1 format). Character.}

\item{ReadAbundMatrix}{Matrix listing samples and read abundance.}
}
\description{
ReadProportions
This function uses the total number of raw reads and BLAST alignment info to determine the proportion of reads from each taxa. This is used for negative control analysis.
}
\examples{
ReadProportions()
}
\keyword{proportions}
