% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetNegativeControl.R
\name{SetNegativeControl}
\alias{SetNegativeControl}
\title{SetNegativeControl
This function is used to find the proportion of all taxa present in the negative controls.}
\usage{
SetNegativeControl(sample.vector, raw.data.path, tblat.path)
}
\arguments{
\item{sample.vector}{Vector of sample names. Character vector.}

\item{raw.data.path}{Pathway to find zipped fastq files. Character.}

\item{tblat.path}{Pathway to alignment files (tblat.1 format). Character.}
}
\description{
SetNegativeControl
This function is used to find the proportion of all taxa present in the negative controls.
}
\examples{
SetNegativeControl()
}
\keyword{negatives}
